# docs/swagger.yaml
definitions:
...
  CreateRepairVO:
    properties:
      arriveTime:
        description: 到达时间开始
        type: string
      auditorEngine:
        description: 审核-维修工程师
        type: string
      auditorProduct:
        description: 审核-生产组长
        type: string
      confirmor:
        description: 确定人
        type: string
      faultDesc:
        description: 故障描述
        type: string
      finishTime:
        description: 完成时间开始
        type: string
      improveContent:
        description: 永久改善内容
        type: string
      lastTime:
        description: 维修时长(分钟)
        minimum: 0
        type: integer
      lineLevel:
        description: 线别
        type: string
      lockor:
        description: 锁定人
        type: string
      moldCode:
        description: 模具编号
        type: string
      operator:
        description: 操作者
        type: string
      repairContent:
        description: 维修内容
        type: string
      repairLevel:
        description: 维修等级
        type: string
      repairSpares:
        description: 维修备件更换
        items:
          $ref: '#/definitions/RepairSpare'
        type: array
      repairStations:
        description: 维修工站
        items:
          type: string
        type: array
      repairtor:
        description: 维修者
        type: string
      reportTime:
        description: 报告时间开始
        type: string
      unlockor:
        description: 解锁人
        type: string
    type: object
...
  DeleteVO:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    type: object
...
  MoldExceptionVO:
    properties:
      cnt:
        description: 累计次数
        type: integer
      cost:
        description: 累计损失工时，单位分钟
        type: integer
      faultDesc:
        description: 故障描述
        type: string
    type: object
...
  PageProductResumeOutputVO:
    properties:
      completeTime:
        description: 生产完成时间
        type: string
      count:
        description: 工单数量
        type: integer
      id:
        type: integer
      lineLevel:
        description: 线别
        type: string
      moldCode:
        description: 模具编码
        type: string
      moldName:
        description: 模具名称
        type: string
      moldType:
        description: 模具类型
        type: string
      orderCode:
        description: 工单号
        type: string
      partCode:
        description: 零件号
        type: string
      sumCount:
        description: 总计数
        type: integer
    type: object
...
  PageRepairOutVO:
    properties:
      arriveTime:
        description: 到达时间开始
        type: string
      auditorEngine:
        description: 审核-维修工程师
        type: string
      auditorProduct:
        description: 审核-生产组长
        type: string
      code:
        description: 维修编号
        type: string
      confirmor:
        description: 确定人
        type: string
      faultDesc:
        description: 故障描述
        type: string
      finishTime:
        description: 完成时间开始
        type: string
      id:
        type: integer
      improveContent:
        description: 永久改善内容
        type: string
      lastTime:
        description: 维修时长(分钟)
        type: integer
      lineLevel:
        description: 线别
        type: string
      lockor:
        description: 锁定人
        type: string
      moldCode:
        description: 模具编号
        type: string
      operator:
        description: 操作者
        type: string
      partCodeStr:
        description: 模具零件号, '/'拼接
        type: string
      partCodes:
        description: 模具零件号
        items:
          type: string
        type: array
      repairContent:
        description: 维修内容
        type: string
      repairLevel:
        description: 维修等级
        type: string
      repairSpares:
        description: 维修备件
        items:
          $ref: '#/definitions/RepairSpare'
        type: array
      repairStation:
        description: 维修工站
        type: string
      repairStations:
        description: 维修工站
        items:
          type: string
        type: array
      repairtor:
        description: 维修者
        type: string
      reportTime:
        description: 报告时间开始
        type: string
      sumLastTime:
        description: 总维修时长(分钟)
        type: integer
      unlockor:
        description: 解锁人
        type: string
    type: object
...
  RepairSpare:
    properties:
      code:
        type: string
      count:
        type: integer
    type: object
...
  UpdateRepairVO:
    properties:
      arriveTime:
        description: 到达时间开始
        type: string
      auditorEngine:
        description: 审核-维修工程师
        type: string
      auditorProduct:
        description: 审核-生产组长
        type: string
      confirmor:
        description: 确定人
        type: string
      faultDesc:
        description: 故障描述
        type: string
      finishTime:
        description: 完成时间开始
        type: string
      id:
        type: integer
      improveContent:
        description: 永久改善内容
        type: string
      lastTime:
        description: 维修时长(分钟)
        minimum: 0
        type: integer
      lineLevel:
        description: 线别
        type: string
      lockor:
        description: 锁定人
        type: string
      moldCode:
        description: 模具编号
        type: string
      operator:
        description: 操作者
        type: string
      repairContent:
        description: 维修内容
        type: string
      repairLevel:
        description: 维修等级
        type: string
      repairSpares:
        description: 维修备件更换
        items:
          $ref: '#/definitions/RepairSpare'
        type: array
      repairStations:
        description: 维修工站
        items:
          type: string
        type: array
      repairtor:
        description: 维修者
        type: string
      reportTime:
        description: 报告时间开始
        type: string
      unlockor:
        description: 解锁人
        type: string
    type: object
...
info:
  contact: {}
  description: 模具管理API
  title: 模具管理API
  version: "1.0"
paths:
...
  /mold/productresume/excel:
    get:
      consumes:
      - application/json
      parameters:
      - description: 指定ID导出备件
        in: query
        name: ids
        type: array
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 导出
      tags:
      - 生产履历
  /mold/productresume/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 模具编号/零件号/工单号模糊查找
        in: query
        name: codeOrName
        type: string
      - description: 当前页
        in: query
        name: currentPage
        type: integer
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - description: 线别
        in: query
        name: lineLevel
        type: string
      - description: RFID TID
        in: query
        name: rfidTid
        type: string
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageProductResumeOutputVO'
      summary: 生产履历分页
      tags:
      - 生产履历
...
  /mold/repair:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/DeleteVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 删除模具维修
      tags:
      - 模具维修
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateRepairVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 添加模具维修
      tags:
      - 模具维修
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/UpdateRepairVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 更新模具维修
      tags:
      - 模具维修
  /mold/repair/one:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageRepairOutVO'
      summary: 模具维修查看
      tags:
      - 模具维修
  /mold/repair/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 到达时间开始
        in: query
        name: arriveTimeBegin
        type: string
      - description: 到达时间结束
        in: query
        name: arriveTimeEnd
        type: string
      - description: 审核-维修工程师
        in: query
        name: auditorEngine
        type: string
      - description: 审核-生产组长
        in: query
        name: auditorProduct
        type: string
      - description: 维修编号
        in: query
        name: code
        type: string
      - description: 维修编码模糊
        in: query
        name: codeLike
        type: string
      - description: 模具编号和模具名称模糊查找，该参数有值时忽略其它条件参数
        in: query
        name: codeOrName
        type: string
      - description: 确定人
        in: query
        name: confirmor
        type: string
      - description: 当前页
        in: query
        name: currentPage
        type: integer
      - description: 故障描述
        in: query
        name: faultDesc
        type: string
      - description: 完成时间开始
        in: query
        name: finishTimeBegin
        type: string
      - description: 完成时间结束
        in: query
        name: finishTimeEnd
        type: string
      - in: query
        name: id
        type: integer
      - description: 永久改善内容
        in: query
        name: improveContent
        type: string
      - description: 线别
        in: query
        name: lineLevel
        type: string
      - description: 锁定人
        in: query
        name: lockor
        type: string
      - description: 模具编号
        in: query
        name: moldCode
        type: string
      - description: 操作者
        in: query
        name: operator
        type: string
      - description: 模具零件号
        in: query
        name: partCodes
        type: string
      - description: 维修内容
        in: query
        name: repairContent
        type: string
      - description: 维修等级
        in: query
        name: repairLevel
        type: string
      - description: 维修工站
        in: query
        name: repairStations
        type: string
      - description: 维修者
        in: query
        name: repairtor
        type: string
      - description: 报告时间开始
        in: query
        name: reportTimeBegin
        type: string
      - description: 报告时间结束
        in: query
        name: reportTimeEnd
        type: string
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: 解锁人
        in: query
        name: unlockor
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageRepairOutVO'
      summary: 模具维修分页
      tags:
      - 模具维修
...
  /mold/statistics/repair:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MoldExceptionVO'
      summary: 模具异常汇总
      tags:
      - 首页
...
swagger: "2.0"